finish broken use case sample
works fine on 8.5.5.3
but broken on 8.5.5.5

Please note that we need the somehow ugly ibm specific binding xmls
amd web.xml entry because otherwise WAS doesn't expose the webservice at all.
I did not really track down which of those files we really need though
but just reproduced the issue we have in production.